Refactor active-filter.js into generic-list
